Refactor run profiles
