+++ rider/src/main/kotlin/com/jetbrains/rider/plugins/unity/run/configurations/UnityAttachToEditorProfileState.kt
@@ -2,6 +2,7 @@ package com.jetbrains.rider.plugins.unity.run.configurations
 
 import com.intellij.execution.ExecutionResult
 import com.intellij.execution.Executor
+import com.intellij.execution.configurations.RunProfileState
 import com.intellij.execution.runners.ExecutionEnvironment
 import com.intellij.execution.runners.ProgramRunner
 import com.intellij.openapi.diagnostic.Logger
@@ -16,26 +17,37 @@ import com.jetbrains.rider.debugger.DebuggerInitializingState
 import com.jetbrains.rider.debugger.DebuggerWorkerProcessHandler
 import com.jetbrains.rider.debugger.RiderDebugActiveDotNetSessionsTracker
 import com.jetbrains.rider.model.unity.frontendBackend.frontendBackendModel
-import com.jetbrains.rider.plugins.unity.run.UnityDebuggerOutputListener
 import com.jetbrains.rider.plugins.unity.run.configurations.unityExe.UnityExeDebugProfileState
 import com.jetbrains.rider.projectView.solution
-import com.jetbrains.rider.run.IDebuggerOutputListener
 import com.jetbrains.rider.run.WorkerRunInfo
-import com.jetbrains.rider.run.configurations.remote.MonoConnectRemoteProfileState
-import com.jetbrains.rider.util.NetUtils
 
-class UnityAttachToEditorProfileState(private val exeDebugProfileState : UnityExeDebugProfileState, private val remoteConfiguration: UnityAttachToEditorRunConfiguration, executionEnvironment: ExecutionEnvironment)
-    : MonoConnectRemoteProfileState(remoteConfiguration, executionEnvironment) {
+/**
+ * [RunProfileState] to attach to the current Unity editor, optionally entering play mode.
+ *
+ * This will use the passed [UnityExeDebugProfileState] to start the Editor if it's not already running.
+ */
+class UnityAttachToEditorProfileState(private val exeDebugProfileState : UnityExeDebugProfileState,
+                                      private val remoteConfiguration: UnityAttachToEditorRunConfiguration,
+                                      executionEnvironment: ExecutionEnvironment)
+    : UnityAttachProfileState(remoteConfiguration, executionEnvironment, "Unity Editor", true) {
+
     private val logger = Logger.getInstance(UnityAttachToEditorProfileState::class.java)
     private val project = executionEnvironment.project
 
-    override fun execute(executor: Executor?, runner: ProgramRunner<*>): ExecutionResult? {
-        throw UnsupportedOperationException("Should use overload with session")
+    override suspend fun createWorkerRunInfo(lifetime: Lifetime, helper: DebuggerHelperHost, port: Int): WorkerRunInfo {
+        // Make sure we have a pid and port to connect. This requires fetching the OS process list, which must happen
+        // on the background thread
+        return withLongBackgroundContext(lifetime) {
+            if (!remoteConfiguration.updatePidAndPort()) {
+                logger.trace("Have not found Unity, would start a new Unity Editor instead.")
+                exeDebugProfileState.createWorkerRunInfo(lifetime, helper, port)
+            }
+            else {
+                super.createWorkerRunInfo(lifetime, helper, port)
+            }
+        }
     }
 
-    override fun execute(executor: Executor, runner: ProgramRunner<*>, workerProcessHandler: DebuggerWorkerProcessHandler): ExecutionResult {
-        throw UnsupportedOperationException("Should use overload with session")
-    }
     override fun execute(executor: Executor, runner: ProgramRunner<*>, workerProcessHandler: DebuggerWorkerProcessHandler, lifetime: Lifetime): ExecutionResult {
         if (remoteConfiguration.play) {
             val lt = lifetime.createNested().lifetime
@@ -68,28 +80,11 @@ class UnityAttachToEditorProfileState(private val exeDebugProfileState : UnityEx
             }
         }
 
-        if (remoteConfiguration.listenPortForConnections)
+        // We couldn't find a running instance, start a new one
+        if (remoteConfiguration.pid == null) {
             return exeDebugProfileState.execute(executor, runner, workerProcessHandler, lifetime)
-        return super.execute(executor, runner, workerProcessHandler)
-    }
-
-    override suspend fun createWorkerRunInfo(lifetime: Lifetime, helper: DebuggerHelperHost, port: Int): WorkerRunInfo {
-        withLongBackgroundContext(lifetime) {
-            if (!remoteConfiguration.updatePidAndPort()) {
-                logger.trace("Have not found Unity, would start a new Unity Editor instead.")
-
-                remoteConfiguration.listenPortForConnections = true
-                remoteConfiguration.port = NetUtils.findFreePort(500013, setOf(port))
-                remoteConfiguration.address = "127.0.0.1"
-            }
         }
 
-        val cmd = super.createWorkerRunInfo(lifetime, helper, port)
-        cmd.commandLine.withUnityExtensionsEnabledEnvironment(project)
-        return cmd
-    }
-
-    override fun getDebuggerOutputEventsListener(): IDebuggerOutputListener {
-        return UnityDebuggerOutputListener(project, remoteConfiguration.address, "Unity Editor", true)
+        return super.execute(executor, runner, workerProcessHandler)
     }
 }
