+++ rider/src/main/kotlin/com/jetbrains/rider/plugins/unity/run/configurations/unityExe/UnityExeDebugProfileState.kt
@@ -3,6 +3,7 @@ package com.jetbrains.rider.plugins.unity.run.configurations.unityExe
 import com.intellij.execution.ExecutionResult
 import com.intellij.execution.Executor
 import com.intellij.execution.configurations.GeneralCommandLine
+import com.intellij.execution.configurations.RunProfileState
 import com.intellij.execution.process.KillableProcessHandler
 import com.intellij.execution.process.ProcessAdapter
 import com.intellij.execution.process.ProcessEvent
@@ -10,7 +11,6 @@ import com.intellij.execution.process.ProcessListener
 import com.intellij.execution.runners.ExecutionEnvironment
 import com.intellij.execution.runners.ProgramRunner
 import com.intellij.openapi.util.Key
-import com.jetbrains.rd.util.addUnique
 import com.jetbrains.rd.util.lifetime.Lifetime
 import com.jetbrains.rd.util.lifetime.onTermination
 import com.jetbrains.rider.debugger.DebuggerHelperHost
@@ -19,32 +19,38 @@ import com.jetbrains.rider.debugger.tryWriteMessageToConsoleView
 import com.jetbrains.rider.model.debuggerWorker.OutputMessageWithSubject
 import com.jetbrains.rider.model.debuggerWorker.OutputSubject
 import com.jetbrains.rider.model.debuggerWorker.OutputType
+import com.jetbrains.rider.plugins.unity.run.configurations.UnityAttachProfileState
 import com.jetbrains.rider.plugins.unity.run.configurations.withUnityExtensionsEnabledEnvironment
 import com.jetbrains.rider.run.ExternalConsoleMediator
 import com.jetbrains.rider.run.WorkerRunInfo
-import com.jetbrains.rider.run.configurations.remote.MonoConnectRemoteProfileState
 import com.jetbrains.rider.run.configurations.remote.RemoteConfiguration
 import com.jetbrains.rider.run.createEmptyConsoleCommandLine
 import com.jetbrains.rider.run.withRawParameters
 import com.jetbrains.rider.util.NetUtils
 
-class UnityExeDebugProfileState(private val exeConfiguration : UnityExeConfiguration, private val remoteConfiguration: RemoteConfiguration,
-                                executionEnvironment: ExecutionEnvironment)
-    : MonoConnectRemoteProfileState(remoteConfiguration, executionEnvironment) {
+/**
+ * [RunProfileState] to launch a Unity executable (player or editor) and attach the debugger
+ */
+class UnityExeDebugProfileState(private val exeConfiguration : UnityExeConfiguration,
+                                private val remoteConfiguration: RemoteConfiguration,
+                                executionEnvironment: ExecutionEnvironment,
+                                isEditor: Boolean = false)
+    : UnityAttachProfileState(remoteConfiguration, executionEnvironment, "Unity Executable", isEditor) {
 
-    override fun execute(executor: Executor?, runner: ProgramRunner<*>): ExecutionResult? {
-        throw UnsupportedOperationException("Should use overload with session")
-    }
+    override suspend fun createWorkerRunInfo(lifetime: Lifetime, helper: DebuggerHelperHost, port: Int): WorkerRunInfo {
+        val runCmd = super.createWorkerRunInfo(lifetime, helper, port)
 
-    override fun execute(executor: Executor, runner: ProgramRunner<*>, workerProcessHandler: DebuggerWorkerProcessHandler): ExecutionResult {
-        throw UnsupportedOperationException("Should use overload with session")
+        remoteConfiguration.listenPortForConnections = true
+        remoteConfiguration.port = NetUtils.findFreePort(500013, setOf(port))
+        remoteConfiguration.address = "127.0.0.1"
+
+        return runCmd
     }
 
     override fun execute(executor: Executor, runner: ProgramRunner<*>, workerProcessHandler: DebuggerWorkerProcessHandler, lifetime: Lifetime): ExecutionResult {
-        val envs = exeConfiguration.parameters.envs.toMutableMap()
-        envs.addUnique(lifetime, "MONO_ARGUMENTS", "--debugger-agent=transport=dt_socket,address=127.0.0.1:${remoteConfiguration.port},server=n,suspend=y")
         val runCommandLine = createEmptyConsoleCommandLine(exeConfiguration.parameters.useExternalConsole)
-            .withEnvironment(envs)
+            .withEnvironment(exeConfiguration.parameters.envs)
+            .withEnvironment("MONO_ARGUMENTS", "--debugger-agent=transport=dt_socket,address=127.0.0.1:${remoteConfiguration.port},server=n,suspend=y")
             .withUnityExtensionsEnabledEnvironment(executionEnvironment.project)
             .withParentEnvironmentType(if (exeConfiguration.parameters.isPassParentEnvs) {
                 GeneralCommandLine.ParentEnvironmentType.CONSOLE
@@ -56,7 +62,11 @@ class UnityExeDebugProfileState(private val exeConfiguration : UnityExeConfigura
             .withRawParameters(exeConfiguration.parameters.programParameters)
 
         val commandLineString = runCommandLine.commandLineString
+
         val monoConnectResult = super.execute(executor, runner, workerProcessHandler)
+
+        // Once the frontend starts the debugger worker process, we'll start the Unity exe, and terminate it when the
+        // debug session ends
         workerProcessHandler.debuggerWorkerRealHandler.addProcessListener(object : ProcessAdapter() {
             override fun startNotified(event: ProcessEvent) {
                 val targetProcessHandler = if (exeConfiguration.parameters.useExternalConsole)
@@ -92,15 +102,5 @@ class UnityExeDebugProfileState(private val exeConfiguration : UnityExeConfigura
         })
 
         return monoConnectResult
-    }
-
-    override suspend fun createWorkerRunInfo(lifetime: Lifetime, helper: DebuggerHelperHost, port: Int): WorkerRunInfo {
-        val runCmd = super.createWorkerRunInfo(lifetime, helper, port)
-
-        remoteConfiguration.listenPortForConnections = true
-        remoteConfiguration.port = NetUtils.findFreePort(500013, setOf(port))
-        remoteConfiguration.address = "127.0.0.1"
-
-        return runCmd
     }
 }
